+++ src/surface_nets.jl
@@ -15,7 +15,7 @@ include("lut/sn.jl")
 Generate a mesh using naive surface nets.
 This takes the center of mass of the voxel as the vertex for each cube.
 """
-function isosurface(sdf::AbstractArray{T, 3}, method::NaiveSurfaceNets, ::Type{VertType}=SVector{3,Float32}, ::Type{FaceType}=SVector{4, Int};
+function isosurface(sdf::AbstractArray{T, 3}, method::NaiveSurfaceNets, ::Type{VertType}=SVector{3,Float64}, ::Type{FaceType}=SVector{4, Int};
                     origin=VertType(-1,-1,-1), widths=VertType(2,2,2)) where {T, VertType, FaceType}
 
     scale = widths ./ VertType(size(sdf) .- 1)  # subtract 1 because an SDF with N points per side has N-1 cells
@@ -95,7 +95,7 @@ Generate a mesh using naive surface nets.
 This takes the center of mass of the voxel as the vertex for each cube.
 """
 function isosurface(f::Function, method::NaiveSurfaceNets,
-                    ::Type{VertType}=SVector{3,Float32}, ::Type{FaceType}=SVector{4, Int};
+                    ::Type{VertType}=SVector{3,Float64}, ::Type{FaceType}=SVector{4, Int};
                     origin=VertType(-1,-1,-1), widths=VertType(2,2,2),
                     samples::NTuple{3,T}=_DEFAULT_SAMPLES) where {T <: Integer, VertType, FaceType}
